+++ odoo/local-src/delivery_rounds/models/round_instance.py
@@ -5,6 +5,7 @@
 from ast import literal_eval
 import math
 from datetime import datetime
+from itertools import groupby
 
 from odoo import api, fields, models, _
 from odoo.exceptions import Warning as UserError
@@ -96,13 +97,12 @@ class RoundInstance(models.Model):
     picking_ids = fields.One2many(
         'stock.picking', 'delivery_round_id', 'Pickings',
         domain=[('picking_type_subcode', '=', 'PICK')],
-        states={'done': [('readonly', True)]},
+        readonly=True,
         )
     shipping_ids = fields.One2many(
         'stock.picking', 'delivery_round_id', 'Deliveries',
         domain=[('picking_type_code', '=', 'outgoing')],
-        states={'done': [('readonly', True)]},
-        # readonly=True,
+        readonly=True,
         )
 
     complete_name = fields.Char(
@@ -185,17 +185,21 @@ class RoundInstance(models.Model):
 
         self.itinerary_ids |= itineraries
 
-        partner_ids = itineraries.mapped('partner_position_ids.partner_id.id')
+        partners = itineraries.mapped('partner_position_ids.partner_id')
+        # Itineraries can contain partner of any type. So extract corresponding
+        # delivery address
+        partners_delivery_ids = [partner.address_get(['delivery'])['delivery']
+                                 for partner in partners]
 
         picking_confirmed = self.env['stock.picking'].search([
             ('delivery_round_id', '=', False),
-            ('partner_id', 'in', partner_ids),
+            ('partner_id', 'in', partners_delivery_ids),
             ('state', '=', 'confirmed')])
         self._assign_pickings(picking_confirmed)
 
     def _assign_pickings(self, pickings, no_prepare=False):
         self.ensure_one()
-        _logger.debug("Assign to delivery round %s the pickings %s",
+        _logger.debug("Assign to round instance %s the pickings %s",
                       self.id, pickings.ids)
 
         pickings.filtered(
@@ -214,12 +218,19 @@ class RoundInstance(models.Model):
             ('id', 'in', pickings.ids),
             ('state', 'in', ('partially_available', 'assigned'))])
         if pickings_assigned:
-            _logger.debug("Add/Propagate to delivery round %s the pickings %s",
-                          self.id, pickings.ids)
-            # Note: a constrain on picking.delivery_round_id will propagate to
-            # group and set the rank
-            pickings_assigned.with_context(round_assigned=True).write({
-                'delivery_round_id': self.id})
+            def key(r):
+                partner = r.partner_id
+                # If delivery address is a contact, take parent
+                if partner.type == 'contact' and partner.parent_id:
+                    partner = partner.parent_id
+                return partner
+
+            for partner, pickings_bypartner_iter in groupby(
+                    pickings_assigned.sorted(key=key), key=key):
+                ric = self._add_customer(partner)
+                pickings_bypartner = reduce(
+                    lambda x, y: x | y, pickings_bypartner_iter)
+                ric._link_pickings(pickings_bypartner)
         return pickings_assigned
 
     @api.multi
@@ -228,7 +239,8 @@ class RoundInstance(models.Model):
         customer.ensure_one()
         ric = self.env['round.instance.customer'].search([
             ('delivery_round_id', '=', self.id),
-            ('partner_id', '=', customer.id)])
+            ('partner_id', '=', customer.id),
+            ('delivered', '!=', True)])
         rank = 0
         if not ric:
             pos = self.env['round.itinerary.position'].search([
@@ -236,29 +248,13 @@ class RoundInstance(models.Model):
                 ('partner_id', '=', customer.id)])
             if pos:
                 rank = (pos.sequence + pos.itinerary_id.sequence*1000) * 1000
-            self.env['round.instance.customer'].sudo().create({
+            _logger.warn("Partner added on delivery %s", self.id)
+            ric = self.env['round.instance.customer'].sudo().create({
                 'delivery_round_id': self.id,
                 'partner_id': customer.id,
                 'rank': rank,
                 })
-        else:
-            rank = ric.rank
-        return rank
-
-    @api.multi
-    def _remove_customer(self, customer):
-        self.ensure_one()
-        if not self.env['stock.picking'].search([
-                ('delivery_round_id', '=', self.id),
-                ('partner_id', '=', customer.id),
-                ('state', '!=', 'cancel'),
-                ]):
-            ric = self.env['round.instance.customer'].search([
-                ('delivery_round_id', '=', self.id),
-                ('partner_id', '=', customer.id),
-                ])
-            if ric:
-                ric.sudo().unlink()
+        return ric
 
     @api.model
     def find(self, partner):
@@ -407,23 +403,17 @@ class RoundInstance(models.Model):
             'stat_time_picking': time_now(self),
             })
 
-    @api.one
+    @api.multi
     def button_deliver(self):
-        """ Validate all deliveries that are available. Mark as done and unlink
-        other deliveries """
-        for shipping in self.shipping_ids:
-            if shipping.state in ('assigned', 'partially_available'):
-                for pack in shipping.pack_operation_ids:
-                    if pack.product_qty > 0:
-                        pack.qty_done = pack.product_qty
-                        for plot in pack.pack_lot_ids:
-                            if plot.qty_todo > 0:
-                                plot.qty = plot.qty_todo
-                    else:
-                        pack.unlink()
-                shipping.do_transfer()
-            elif shipping.state == 'confirmed':
-                shipping.delivery_round_id = None
+        """ Deliver all customers. This validates all shipping orders that are
+        available.
+        Mark as done and unlink other deliveries
+        """
+        icust = self.mapped('instance_customer_ids')
+        icust.filtered(lambda c: not c.delivered)._deliver()
+        # Remove all customer where no picking have been processed
+        icust._remove()
+        # Close delivery round
         self.button_done()
 
     @api.multi
@@ -525,16 +515,8 @@ class RoundInstance(models.Model):
 
 class RoundInstanceCustomer(models.Model):
     _name = 'round.instance.customer'
-    _order = 'rank'
+    _order = 'rank,delivered,write_date desc'
     _rec_name = 'partner_id'
-
-    _sql_constraints = [
-        (
-            'unique_instance_partner',
-            'UNIQUE(delivery_round_id, partner_id)',
-            _('The customer must be unique in a delivery round.')
-        ),
-    ]
 
     delivery_round_id = fields.Many2one(
         comodel_name='round.instance',
@@ -559,6 +541,48 @@ class RoundInstanceCustomer(models.Model):
         string='Rank',
     )
 
+    picking_ids = fields.One2many(
+        'stock.picking', 'delivery_round_customer_id', 'Pickings',
+        readonly=True)
+
+    @api.multi
+    def _link_pickings(self, pickings):
+        self.ensure_one()
+        # Link all pickings/shippings
+        shippings = pickings._get_all_dest_pickings().filtered(
+            lambda r: r.picking_type_code == 'outgoing')
+        # ensure all related pickings are assigned to the same delivery
+        # round
+        # Use | to let it work in tests with one step delivery
+        pickings |= shippings._get_all_src_pickings()
+        # Note that in our case, an open picking cannot have multiple open
+        # shippings, so we don't have to ensure a picking is not already done
+        # for another delivery round
+        pickings = pickings.filtered(
+            lambda r: r.state in (
+                'waiting',
+                'confirmed',
+                'partially_available',
+                'assigned') and
+            r.delivery_round_customer_id.id != self.id)
+        if pickings:
+            _logger.debug("Link to delivery round the pickings/shippings %s",
+                          pickings)
+            pickings.with_context(noround_write=True).write({
+                'delivery_round_customer_id': self.id,
+                'rank': self.rank})
+
+    def _remove(self):
+        """ Remvove partner from round instance if no more pickings or all
+        canceled """
+        if not self.mapped('picking_ids').filtered(
+                lambda p: p.state != 'cancel'):
+            _logger.debug(
+                "Removing customers %s from round instance %s",
+                self.mapped('partner_id').ids,
+                self.mapped('delivery_round_id').ids)
+            self.unlink()
+
     @api.multi
     @api.constrains('rank')
     def _propagate_rank(self):
@@ -566,18 +590,66 @@ class RoundInstanceCustomer(models.Model):
             rank = instance_customer.rank
             # when we set a rank on a round instance customer,
             # we copy that value on the pickings
-            pickings = self.delivery_round_id.shipping_ids.filtered(
-                lambda p:
-                p.partner_id == instance_customer.partner_id and
-                p.rank != rank
-            )
-            pickings += self.delivery_round_id.picking_ids.filtered(
-                lambda p:
-                p.partner_id == instance_customer.partner_id and
-                p.rank != rank
-            )
+            pickings = self.picking_ids.filtered(lambda p: p.rank != rank)
+            if not pickings:
+                continue
             _logger.debug(
                 "Rank set on round instance customer %s. Propagate to "
                 "pickings and shippings %s",
                 self.ids, pickings.ids)
             pickings.write({'rank': rank})
+
+    count_picking_progress = fields.Char(
+        'Picking Progress',
+        compute='_get_count_picking',
+        readonly=True)
+
+    @api.depends('picking_ids')
+    def _get_count_picking(self):
+        for rec in self:
+            pickings = rec.picking_ids.filtered(
+                lambda r: r.picking_type_subcode == 'PICK')
+            count_done = len(pickings.filtered(
+                lambda r: r.state == ('done')))
+            count_total = len(pickings.filtered(
+                lambda r: r.state in ('partially_available', 'assigned',
+                                      'done')))
+            rec.count_picking_progress = '%s/%s' % (count_done, count_total)
+
+    delivered = fields.Boolean('Delivered')
+
+    def button_deliver(self):
+        """ Validate all shipping orders that are available """
+        self.ensure_one()
+        self._deliver()
+        if not self.picking_ids:
+            # Nothing was picked, all pickings have been disconnected
+            raise UserError(_("No picking have been processed yet"))
+
+    def _deliver(self):
+        """ Validate all shipping orders that are available """
+        shippings = self.mapped('picking_ids').filtered(
+            lambda p: p.picking_type_code == 'outgoing')
+        for shipping in shippings.filtered(
+                lambda p: p.state in ('assigned', 'partially_available')):
+            for pack in shipping.pack_operation_ids:
+                if pack.product_qty > 0:
+                    pack.qty_done = pack.product_qty
+                    for plot in pack.pack_lot_ids:
+                        if plot.qty_todo > 0:
+                            plot.qty = plot.qty_todo
+                else:
+                    pack.unlink()
+            shipping.do_transfer()
+        self.mapped('picking_ids').filtered(
+            lambda p: p.state not in ('cancel', 'done')).\
+            write({'delivery_round_customer_id': False})
+        self.write({'delivered': True})
+
+    def print_deliveryslip(self):
+        shippings = self.picking_ids.filtered(
+            lambda p: p.picking_type_code == 'outgoing')
+        shipping_done = shippings.filtered(
+            lambda shipping: shipping.state == 'done')
+        return self.env['report'].get_action(shipping_done,
+                                             'stock.report_deliveryslip')